codeclimate fix
